Refactor by divide `Label` composable function
